Moved common parts of free/open/netbsd code to disc_bsd.c.
