+++ CMakeLists.txt
@@ -72,18 +72,20 @@ ELSE()
 ENDIF()
 
 # choose platform dependent source files
-IF(libdiscid_OS MATCHES "win32")
+IF(libdiscid_OS EQUAL "win32")
     SET(libdiscid_OSDEP_SRCS src/toc.c src/disc_win32.c)
     SET(libdiscid_RCS ${CMAKE_CURRENT_BINARY_DIR}/versioninfo.rc)
-ELSEIF(libdiscid_OS MATCHES "darwin")
-    SET(libdiscid_OSDEP_SRCS src/toc.c src/unix.c src/disc_darwin.c)
-    FIND_LIBRARY(COREFOUNDATION_LIBRARY CoreFoundation)
-    SET(libdiscid_OSDEP_LIBS ${COREFOUNDATION_LIBRARY} IOKit)
-ELSEIF(libdiscid_OS MATCHES "generic")
+ELSEIF(libdiscid_OS EQUAL "generic")
     SET(libdiscid_OSDEP_SRCS src/disc_${libdiscid_OS}.c)
 ELSE()
     # unix platforms are the standard/default case
     SET(libdiscid_OSDEP_SRCS src/toc.c src/unix.c src/disc_${libdiscid_OS}.c)
+    IF(libdiscid_OS MATCHES "bsd$")
+        SET(libdiscid_OSDEP_SRCS ${libdiscid_OSDEP_SRCS} src/disc_bsd.c)
+    ELSEIF(libdiscid_OS EQUAL "darwin")
+        FIND_LIBRARY(COREFOUNDATION_LIBRARY CoreFoundation)
+        SET(libdiscid_OSDEP_LIBS ${COREFOUNDATION_LIBRARY} IOKit)
+    ENDIF()
 ENDIF()
 
 STRING(REPLACE ";" " " libdiscid_OSDEP_STR "${libdiscid_OSDEP_SRCS}")
