+++ Makefile.am
@@ -85,7 +85,7 @@ libdiscid_la_LDFLAGS += -framework CoreFoundation -framework IOKit
 libdiscid_la_SOURCES += src/toc.c src/unix.c src/disc_darwin.c
 endif
 if OS_FREEBSD
-libdiscid_la_SOURCES += src/toc.c src/unix.c src/disc_freebsd.c
+libdiscid_la_SOURCES += src/toc.c src/unix.c src/disc_bsd.c src/disc_freebsd.c
 endif
 if OS_GENERIC
 libdiscid_la_SOURCES += src/disc_generic.c
@@ -94,7 +94,7 @@ if OS_LINUX
 libdiscid_la_SOURCES += src/toc.c src/unix.c src/disc_linux.c
 endif
 if OS_NETBSD
-libdiscid_la_SOURCES += src/toc.c src/unix.c src/disc_netbsd.c
+libdiscid_la_SOURCES += src/toc.c src/unix.c src/disc_bsd.c src/disc_netbsd.c
 endif
 #if OS_QNX
 #libdiscid_la_LIBADD += -lsocket
